1protocol_neuro/requirements.txt
nengo
lava-nc
numpy
matplotlib
